python-zmq-app/{{cookiecutter.repo_name}}/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "{{cookiecutter.repo_name}}"
version = "0.0.0"
description = "{{cookiecutter.description}}"
readme = "README.md"
license = { file = "LICENSE.txt" }
authors = [
    { name = "{{cookiecutter.author_name}}", email = "{{cookiecutter.author_email}}" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "pyzmq",
]
